# api/openapi.yaml
openapi: 3.1.1
info:
  title: Bounty Board Work API
  version: 1.0.0
  description: Versioned work-management API for authenticated people and agents.
security:
  - bearerAuth: []
  - sessionCookie: []
paths:
...
  /api/v1/tasks/{number}:
    parameters:
      - $ref: "#/components/parameters/TaskNumber"
    get:
      operationId: getTask
      summary: Get a task
      responses:
        "200":
          $ref: "#/components/responses/Task"
        default:
          $ref: "#/components/responses/Problem"
    patch:
      operationId: updateTask
      summary: Update a task
      parameters:
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskPatch"
      responses:
        "200":
          $ref: "#/components/responses/Task"
        default:
          $ref: "#/components/responses/Problem"
  /api/v1/tasks/{number}/archive:
    parameters:
      - $ref: "#/components/parameters/TaskNumber"
    post:
      operationId: archiveTask
      summary: Archive a task
      parameters:
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          $ref: "#/components/responses/Task"
        default:
          $ref: "#/components/responses/Problem"
  /api/v1/tasks/{number}/restore:
    parameters:
      - $ref: "#/components/parameters/TaskNumber"
    post:
      operationId: restoreTask
      summary: Restore a task
      parameters:
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          $ref: "#/components/responses/Task"
        default:
          $ref: "#/components/responses/Problem"
  /api/v1/tasks/{number}/comments:
    parameters:
      - $ref: "#/components/parameters/TaskNumber"
    get:
      operationId: listTaskComments
      summary: List task comments
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          $ref: "#/components/responses/CommentList"
        default:
          $ref: "#/components/responses/Problem"
    post:
      operationId: createTaskComment
      summary: Add a task comment
      parameters:
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentWrite"
      responses:
        "201":
          $ref: "#/components/responses/CommentCreated"
        default:
          $ref: "#/components/responses/Problem"
  /api/v1/tasks/{number}/comments/{id}:
    parameters:
      - $ref: "#/components/parameters/TaskNumber"
      - $ref: "#/components/parameters/ResourceID"
    patch:
      operationId: updateTaskComment
      summary: Update a task comment
      parameters:
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentWrite"
      responses:
        "200":
          $ref: "#/components/responses/Comment"
        default:
          $ref: "#/components/responses/Problem"
    delete:
      operationId: deleteTaskComment
      summary: Delete a task comment
      parameters:
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        default:
          $ref: "#/components/responses/Problem"
...
  /api/v1/labels/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceID"
    patch:
      operationId: updateLabel
      summary: Rename a label
      parameters:
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabelWrite"
      responses:
        "200":
          $ref: "#/components/responses/Label"
        default:
          $ref: "#/components/responses/Problem"
    delete:
      operationId: deleteLabel
      summary: Delete a label
      parameters:
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        default:
          $ref: "#/components/responses/Problem"
...
  /api/v1/projects/{number}:
    parameters:
      - $ref: "#/components/parameters/ProjectNumber"
    get:
      operationId: getProject
      summary: Get a project and its work aggregate
      responses:
        "200":
          $ref: "#/components/responses/ProjectDetail"
        default:
          $ref: "#/components/responses/Problem"
    patch:
      operationId: updateProject
      summary: Update a project
      parameters:
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectPatch"
      responses:
        "200":
          $ref: "#/components/responses/Project"
        default:
          $ref: "#/components/responses/Problem"
...
  /api/v1/projects/{number}/milestones:
    parameters:
      - $ref: "#/components/parameters/ProjectNumber"
    post:
      operationId: createMilestone
      summary: Create a milestone
      parameters:
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MilestoneCreate"
      responses:
        "201":
          $ref: "#/components/responses/MilestoneCreated"
        default:
          $ref: "#/components/responses/Problem"
  /api/v1/projects/{number}/milestones/{id}:
    parameters:
      - $ref: "#/components/parameters/ProjectNumber"
      - $ref: "#/components/parameters/ResourceID"
    patch:
      operationId: updateMilestone
      summary: Update a milestone
      parameters:
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/ProjectIfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MilestonePatch"
      responses:
        "200":
          $ref: "#/components/responses/Milestone"
        default:
          $ref: "#/components/responses/Problem"
...
  /api/v1/tasks/{number}/criteria:
    parameters:
      - $ref: "#/components/parameters/TaskNumber"
    get:
      operationId: listTaskCriteria
      summary: List task acceptance criteria
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          $ref: "#/components/responses/CriterionList"
        default:
          $ref: "#/components/responses/Problem"
    post:
      operationId: createTaskCriterion
      summary: Create a task acceptance criterion
      parameters:
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriterionCreate"
      responses:
        "201":
          $ref: "#/components/responses/CriterionCreated"
        default:
          $ref: "#/components/responses/Problem"
  /api/v1/projects/{number}/milestones/{id}/criteria:
    parameters:
      - $ref: "#/components/parameters/ProjectNumber"
      - $ref: "#/components/parameters/ResourceID"
    get:
      operationId: listMilestoneCriteria
      summary: List milestone acceptance criteria
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          $ref: "#/components/responses/CriterionList"
        default:
          $ref: "#/components/responses/Problem"
    post:
      operationId: createMilestoneCriterion
      summary: Create a milestone acceptance criterion
      parameters:
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/ProjectIfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriterionCreate"
      responses:
        "201":
          $ref: "#/components/responses/CriterionCreated"
        default:
          $ref: "#/components/responses/Problem"
  /api/v1/criteria/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceID"
    patch:
      operationId: updateCriterion
      summary: Update an acceptance criterion
      parameters:
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriterionPatch"
      responses:
        "200":
          $ref: "#/components/responses/Criterion"
        default:
          $ref: "#/components/responses/Problem"
    delete:
      operationId: deleteCriterion
      summary: Remove an acceptance criterion
      parameters:
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LifecycleRequest"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        default:
          $ref: "#/components/responses/Problem"
  /api/v1/criteria/{id}/checks:
    parameters:
      - $ref: "#/components/parameters/ResourceID"
    post:
      operationId: createAcceptanceCheck
      summary: Record an immutable acceptance check
      parameters:
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcceptanceCheckCreate"
      responses:
        "201":
          $ref: "#/components/responses/AcceptanceCheckCreated"
        default:
          $ref: "#/components/responses/Problem"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    sessionCookie:
      type: apiKey
      in: cookie
      name: bb_session
  parameters:
    Cursor:
      name: cursor
      in: query
      description: Opaque cursor returned by the previous page.
      schema:
        type: string
        maxLength: 2048
    Limit:
      name: limit
      in: query
      description: Page size. Defaults to 50 and is capped at 200.
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    TaskNumber:
      name: number
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    ProjectNumber:
      name: number
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    ResourceID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: One quoted positive integer resource version, for example `"3"`.
      schema:
        type: string
        pattern: '^"[1-9][0-9]*"$'
    ProjectIfMatch:
      name: X-Project-If-Match
      in: header
      required: true
      description: Current project Etag when an operation changes a milestone and its project.
      schema:
        type: string
        pattern: '^"[1-9][0-9]*"$'
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Required for bearer-authenticated mutations; optional for browser sessions.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: '^[\x21-\x7E]+$'
  headers:
    Etag:
      description: Quoted positive integer resource version.
      schema:
        type: string
        pattern: '^"[1-9][0-9]*"$'
    Location:
      description: Canonical URL of the created resource.
      schema:
        type: string
        format: uri-reference
    RequestID:
      description: Server-owned request correlation ID.
      schema:
        type: string
    IdempotencyReplayed:
      description: Present and true when a stored response was replayed.
      schema:
        type: boolean
    RateLimitLimit:
      description: Sustained per-token request limit per minute.
      schema:
        type: integer
        minimum: 1
    RateLimitRemaining:
      description: Requests currently available in the token bucket.
      schema:
        type: integer
        minimum: 0
    RateLimitReset:
      description: Seconds until the token bucket is full, or until one request is available.
      schema:
        type: integer
        minimum: 0
    RetryAfter:
      description: Seconds before a retry is recommended.
      schema:
        type: integer
        minimum: 1
...
  responses:
...
    Task:
      description: Task
      headers:
        Etag:
          $ref: "#/components/headers/Etag"
        X-Request-Id:
          $ref: "#/components/headers/RequestID"
        Ratelimit-Limit:
          $ref: "#/components/headers/RateLimitLimit"
        Ratelimit-Remaining:
          $ref: "#/components/headers/RateLimitRemaining"
        Ratelimit-Reset:
          $ref: "#/components/headers/RateLimitReset"
        Idempotency-Replayed:
          $ref: "#/components/headers/IdempotencyReplayed"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Task"
...
    CommentList:
      description: Comment collection
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestID"
        Ratelimit-Limit:
          $ref: "#/components/headers/RateLimitLimit"
        Ratelimit-Remaining:
          $ref: "#/components/headers/RateLimitRemaining"
        Ratelimit-Reset:
          $ref: "#/components/headers/RateLimitReset"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CommentList"
    Comment:
      description: Comment
      headers:
        Etag:
          $ref: "#/components/headers/Etag"
        X-Request-Id:
          $ref: "#/components/headers/RequestID"
        Ratelimit-Limit:
          $ref: "#/components/headers/RateLimitLimit"
        Ratelimit-Remaining:
          $ref: "#/components/headers/RateLimitRemaining"
        Ratelimit-Reset:
          $ref: "#/components/headers/RateLimitReset"
        Idempotency-Replayed:
          $ref: "#/components/headers/IdempotencyReplayed"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Comment"
    CommentCreated:
      description: Comment created
      headers:
        Etag:
          $ref: "#/components/headers/Etag"
        Location:
          $ref: "#/components/headers/Location"
        X-Request-Id:
          $ref: "#/components/headers/RequestID"
        Ratelimit-Limit:
          $ref: "#/components/headers/RateLimitLimit"
        Ratelimit-Remaining:
          $ref: "#/components/headers/RateLimitRemaining"
        Ratelimit-Reset:
          $ref: "#/components/headers/RateLimitReset"
        Idempotency-Replayed:
          $ref: "#/components/headers/IdempotencyReplayed"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Comment"
...
    Label:
      description: Label
      headers:
        Etag:
          $ref: "#/components/headers/Etag"
        X-Request-Id:
          $ref: "#/components/headers/RequestID"
        Ratelimit-Limit:
          $ref: "#/components/headers/RateLimitLimit"
        Ratelimit-Remaining:
          $ref: "#/components/headers/RateLimitRemaining"
        Ratelimit-Reset:
          $ref: "#/components/headers/RateLimitReset"
        Idempotency-Replayed:
          $ref: "#/components/headers/IdempotencyReplayed"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Label"
...
    Project:
      description: Project
      headers:
        Etag:
          $ref: "#/components/headers/Etag"
        X-Request-Id:
          $ref: "#/components/headers/RequestID"
        Ratelimit-Limit:
          $ref: "#/components/headers/RateLimitLimit"
        Ratelimit-Remaining:
          $ref: "#/components/headers/RateLimitRemaining"
        Ratelimit-Reset:
          $ref: "#/components/headers/RateLimitReset"
        Idempotency-Replayed:
          $ref: "#/components/headers/IdempotencyReplayed"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Project"
    ProjectDetail:
      description: Project aggregate
      headers:
        Etag:
          $ref: "#/components/headers/Etag"
        X-Request-Id:
          $ref: "#/components/headers/RequestID"
        Ratelimit-Limit:
          $ref: "#/components/headers/RateLimitLimit"
        Ratelimit-Remaining:
          $ref: "#/components/headers/RateLimitRemaining"
        Ratelimit-Reset:
          $ref: "#/components/headers/RateLimitReset"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProjectDetail"
...
    Milestone:
      description: Milestone
      headers:
        Etag:
          $ref: "#/components/headers/Etag"
        X-Request-Id:
          $ref: "#/components/headers/RequestID"
        Ratelimit-Limit:
          $ref: "#/components/headers/RateLimitLimit"
        Ratelimit-Remaining:
          $ref: "#/components/headers/RateLimitRemaining"
        Ratelimit-Reset:
          $ref: "#/components/headers/RateLimitReset"
        Idempotency-Replayed:
          $ref: "#/components/headers/IdempotencyReplayed"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Milestone"
    MilestoneCreated:
      description: Milestone created
      headers:
        Etag:
          $ref: "#/components/headers/Etag"
        Location:
          $ref: "#/components/headers/Location"
        X-Request-Id:
          $ref: "#/components/headers/RequestID"
        Ratelimit-Limit:
          $ref: "#/components/headers/RateLimitLimit"
        Ratelimit-Remaining:
          $ref: "#/components/headers/RateLimitRemaining"
        Ratelimit-Reset:
          $ref: "#/components/headers/RateLimitReset"
        Idempotency-Replayed:
          $ref: "#/components/headers/IdempotencyReplayed"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Milestone"
    CriterionList:
      description: Acceptance criterion collection
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestID"
        Ratelimit-Limit:
          $ref: "#/components/headers/RateLimitLimit"
        Ratelimit-Remaining:
          $ref: "#/components/headers/RateLimitRemaining"
        Ratelimit-Reset:
          $ref: "#/components/headers/RateLimitReset"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CriterionList"
    Criterion:
      description: Acceptance criterion
      headers:
        Etag:
          $ref: "#/components/headers/Etag"
        X-Request-Id:
          $ref: "#/components/headers/RequestID"
        Ratelimit-Limit:
          $ref: "#/components/headers/RateLimitLimit"
        Ratelimit-Remaining:
          $ref: "#/components/headers/RateLimitRemaining"
        Ratelimit-Reset:
          $ref: "#/components/headers/RateLimitReset"
        Idempotency-Replayed:
          $ref: "#/components/headers/IdempotencyReplayed"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AcceptanceCriterion"
    CriterionCreated:
      description: Acceptance criterion created
      headers:
        Etag:
          $ref: "#/components/headers/Etag"
        Location:
          $ref: "#/components/headers/Location"
        X-Request-Id:
          $ref: "#/components/headers/RequestID"
        Ratelimit-Limit:
          $ref: "#/components/headers/RateLimitLimit"
        Ratelimit-Remaining:
          $ref: "#/components/headers/RateLimitRemaining"
        Ratelimit-Reset:
          $ref: "#/components/headers/RateLimitReset"
        Idempotency-Replayed:
          $ref: "#/components/headers/IdempotencyReplayed"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AcceptanceCriterion"
    AcceptanceCheckCreated:
      description: Acceptance check created
      headers:
        Location:
          $ref: "#/components/headers/Location"
        X-Request-Id:
          $ref: "#/components/headers/RequestID"
        Ratelimit-Limit:
          $ref: "#/components/headers/RateLimitLimit"
        Ratelimit-Remaining:
          $ref: "#/components/headers/RateLimitRemaining"
        Ratelimit-Reset:
          $ref: "#/components/headers/RateLimitReset"
        Idempotency-Replayed:
          $ref: "#/components/headers/IdempotencyReplayed"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AcceptanceCheck"
    NoContent:
      description: Mutation completed without a response body
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestID"
        Ratelimit-Limit:
          $ref: "#/components/headers/RateLimitLimit"
        Ratelimit-Remaining:
          $ref: "#/components/headers/RateLimitRemaining"
        Ratelimit-Reset:
          $ref: "#/components/headers/RateLimitReset"
        Idempotency-Replayed:
          $ref: "#/components/headers/IdempotencyReplayed"
    Problem:
      description: Request failed
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestID"
        Ratelimit-Limit:
          $ref: "#/components/headers/RateLimitLimit"
        Ratelimit-Remaining:
          $ref: "#/components/headers/RateLimitRemaining"
        Ratelimit-Reset:
          $ref: "#/components/headers/RateLimitReset"
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    TaskStatus:
      type: string
      enum: [todo, in_progress, in_review, done, cancelled]
    TaskPriority:
      type: string
      enum: [none, low, medium, high, urgent]
...
    NullableUUID:
      oneOf:
        - type: string
          format: uuid
        - type: "null"
    NullableDate:
      oneOf:
        - type: string
          format: date
        - type: "null"
...
    TaskPatch:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
        description:
          type: string
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          $ref: "#/components/schemas/TaskPriority"
        assignee_id:
          $ref: "#/components/schemas/NullableUUID"
        due_date:
          $ref: "#/components/schemas/NullableDate"
        label_ids:
          type: array
          uniqueItems: true
          items:
            type: string
            format: uuid
        project_number:
          type: integer
          format: int64
          minimum: 1
        milestone_id:
          $ref: "#/components/schemas/NullableUUID"
...
    Task:
      type: object
      additionalProperties: false
      required:
        - id
        - number
        - version
        - title
        - description
        - status
        - priority
        - creator
        - project
        - labels
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        number:
          type: integer
          format: int64
          minimum: 1
        version:
          type: integer
          format: int64
          minimum: 1
        title:
          type: string
        description:
          type: string
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          $ref: "#/components/schemas/TaskPriority"
        assignee:
          $ref: "#/components/schemas/UserRef"
        creator:
          $ref: "#/components/schemas/UserRef"
        due_date:
          type: string
          format: date
        project:
          $ref: "#/components/schemas/ProjectRef"
        milestone:
          $ref: "#/components/schemas/MilestoneRef"
        labels:
          type: array
          items:
            $ref: "#/components/schemas/Label"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        archived_at:
          type: string
          format: date-time
...
    CommentWrite:
      type: object
      additionalProperties: false
      required: [body]
      properties:
        body:
          type: string
          minLength: 1
    Comment:
      type: object
      additionalProperties: false
      required: [id, task_id, author_id, body, version, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        task_id:
          type: string
          format: uuid
        author_id:
          type: string
          format: uuid
        body:
          type: string
        version:
          type: integer
          format: int64
          minimum: 1
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CommentList:
      allOf:
        - $ref: "#/components/schemas/Page"
        - type: object
          additionalProperties: false
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Comment"
...
    LabelWrite:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
    Label:
      type: object
      additionalProperties: false
      required: [id, name, version, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        version:
          type: integer
          format: int64
          minimum: 1
        created_at:
          type: string
          format: date-time
...
    ProjectPatch:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 500
        description:
          type: string
        owner_id:
          type: string
          format: uuid
...
    MilestoneCreate:
      type: object
      additionalProperties: false
      required: [name, outcome, owner_id, position]
      properties:
        name:
          type: string
          minLength: 1
        outcome:
          type: string
          minLength: 1
        description:
          type: string
        owner_id:
          type: string
          format: uuid
        target_date:
          $ref: "#/components/schemas/NullableDate"
        position:
          type: integer
          minimum: 0
    MilestonePatch:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
        outcome:
          type: string
          minLength: 1
        description:
          type: string
        owner_id:
          type: string
          format: uuid
        target_date:
          $ref: "#/components/schemas/NullableDate"
        position:
          type: integer
          minimum: 0
...
    CriterionCreate:
      type: object
      additionalProperties: false
      required: [criterion, verification_instructions, position]
      properties:
        criterion:
          type: string
          minLength: 1
        verification_instructions:
          type: string
          minLength: 1
        position:
          type: integer
          minimum: 0
    CriterionPatch:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        criterion:
          type: string
          minLength: 1
        verification_instructions:
          type: string
          minLength: 1
        position:
          type: integer
          minimum: 0
...
    CriterionList:
      allOf:
        - $ref: "#/components/schemas/Page"
        - type: object
          additionalProperties: false
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/AcceptanceCriterion"
    AcceptanceCheckCreate:
      type: object
      additionalProperties: false
      required: [criterion_revision, outcome, evidence]
      properties:
        criterion_revision:
          type: integer
          minimum: 1
        outcome:
          $ref: "#/components/schemas/AcceptanceOutcome"
        evidence:
          type: string
          minLength: 1
...
    ProjectDetail:
      type: object
      additionalProperties: false
      required: [project, milestones, tasks, activity]
      properties:
        project:
          $ref: "#/components/schemas/Project"
        milestones:
          type: array
          items:
            $ref: "#/components/schemas/Milestone"
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/Task"
        activity:
          type: array
          items:
            $ref: "#/components/schemas/ProjectActivity"
    LifecycleRequest:
      type: object
      additionalProperties: false
      properties:
        reason:
          type: string
          maxLength: 2000
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, detail, instance, code, request_id]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          $ref: "#/components/schemas/ProblemCode"
        request_id:
          type: string
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ValidationError"
        current_version:
          type: integer
          format: int64
          minimum: 1
        retryable:
          type: boolean